% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allpayGame.R
\name{allpayGame}
\alias{allpayGame}
\title{Tabulate entry and exit game results.}
\usage{
allpayGame(sheet = NULL, endowment = 5, prize = 4, ...)
}
\arguments{
\item{sheet}{(required) is a character string url corresponding to the Google Sheets location containing the individual submissions.}

\item{endowment}{is the size of the initial endowment of points the instructor wishes to give each student.}

\item{prize}{is the value of the license or prize each interest group (student) is bidding on.}
}
\value{
\code{type} returns the type of activity (allpayGame).

\code{results} returns the original submissions (with market prices and points per round added).

\code{winner} returns the name of the winner and the size of the prize.

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Tabulates and assigns points for the results of a simple in-class entry and exit game using random values.
}
\details{
\code{allpayGame} tabulates the results of a simple lobbying game based on a modified version of Goeree and Holt (1999). By default students receive an "endowment" of "extra credit" points (default is 5) at the beginning of a round of bidding. In one iteration, students "bid" on a monopoly license "prize" that is (default is 4 points) by making an anonymous number of "contributions" to politicians in order to influence the decision by the government. Each contribution costs 1 point, and is tabulated as the number of times a student submits the Google Form. The winner of the auction is determined by a lottery from a random row sampled from the responses.
}
\references{
Goeree, Jacob K. & Charles A. Holt (1999). Classroom Games: Rent-Seeking and the Inefficiency of Non-market Allocations. \emph{Journal of Economic Perspectives,} 13(3), pp.217-226.
}
