% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchoringGame.R
\name{anchoringGame}
\alias{anchoringGame}
\title{Tabulate anchoring game results.}
\usage{
anchoringGame(sheet, ...)
}
\arguments{
\item{sheet}{(required) is a character string url corresponding to the Google Sheets location containing the individual submissions.}
}
\value{
\code{type} returns the type of activity (anchoringGame).

\code{results} returns the original submissions.

\code{SummaryTable} returns the summary table (with difference in means test statistics).
}
\description{
Tabulates and assigns points for the results of a simple in-class anchoring game.
}
\details{
\code{anchoringGame} tabulates the results of a simple anchoring game based on the classroom adaptation of Tversky and Kahneman (1974) presented in Gelman and Glickman (2000). Students form receive a random signal (10 or 65 with probabilities {0.5, 0.5}) from a Google Sheet. They enter their name and their signal in a Google Form. Then they are asked to make two guesses about an unknown quantity: in the original example the unknown quantity is the percentage of countries in the world that are located on the continent of Africa. The first question is whether they believe this quantity is higher or lower than their signal. The second question asks them to guess the quantity. This function tabulates the guesses by group in a \code{gtsummay} table with the corresponding p-value for the difference-in-means test. Plotting the output produces a box plot of the responses by group.
}
\references{
Gelman, A. and Nolan, D., 2002. Some Class-Participation Demonstrations for Introductory Probability and Statistics \emph{Journal of Educational and Behavioral Statistics.} 25(1): 84-100.
Tversky, A. and Kahneman, D., 1974. Judgement under Uncertainty: Heuristics and Biases \emph{Science, New Series.} 1985(4175): 1124-1131.
}
