% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertrandGame.R
\name{bertrandGame}
\alias{bertrandGame}
\title{Tabulate results for a simple in-class stag hunt game.}
\usage{
bertrandGame(
  sheet,
  a = 10,
  b = -1,
  c = 6,
  f = 0,
  partners = "random",
  names = NULL,
  ...
)
}
\arguments{
\item{sheet}{(required) is a character string sheet ID corresponding to the Google Sheets location containing the individual submissions.}

\item{a}{is the value of the intercept of the linear inverse-demand function (default is 10).}

\item{b}{is the value of the slope of the linear inverse-demand function (default is -1).}

\item{c}{is the value of the firm's marginal cost (default is 6).}

\item{f}{is the value of the firm's fixed cost (default is 0).}

\item{partners}{is a character string equal to \code{"students"} or \code{'random'} indicating whether students choose their own partners (default) or whether the function should generate them randomly after the students have decided their strategies. If \code{partners = "students"}, then the form must include fields for the student's own first and last names and fields for their partners' first and last names. If partners are random, then the columns in \code{sheet} should either contain columns named "First Name" and "Last Name", or the user needs to specify the columns in \code{sheet} corresponding to this information to pass to \code{randomGroups()}.}
}
\value{
\code{type} returns the type of activity (\code{bertrandGame}).

\code{results} returns the original submissions (with equilibria and points per round added).

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Tabulate results for a simple in-class stag hunt game.
}
\details{
\code{bertrandGame} tabulates the results of a simple stag hunt game in which students' points depend on their strategy and the strategy chosen by their randomly-assigned partners.
}
