% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibriumGame.R
\name{equilibriumGame}
\alias{equilibriumGame}
\title{Tabulate market equilibrium game results.}
\usage{
equilibriumGame(sheet, auth = FALSE, email = NULL, ...)
}
\arguments{
\item{sheet}{(required) is a character string url corresponding to the Google Sheets location containing the individual submissions.}

\item{auth}{is a logical indicating whether to use an authentication token to access the Sheet containing the individual submissions.}

\item{email}{is an email address that matches the user account containing the Sheet with the individual submissions.}
}
\value{
\code{type} returns the type of activity (equlibriumGame).

\code{results} returns the original submissions (with market price and points added).

\code{schedules} returns a list containing the supply and demand schedules for each round.

\code{equilibria} returns a list containing the equilibria for each round.

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Tabulates and assigns points for the results of a simple in-class market equilibrium game using random values.
}
\details{
\code{equilibriumGame} tabulates the results of a simple equilibrium game based on Holt (1996). The instructor informs the students that they own a single unit of a eCoin currency that each of them values differently. Students then receive a random value from 1 to 10 using a link to a Google Sheet. This number represents their (constant) value they attach to the unit they presently own and for if they were to acquire one more unit of eCoin. Students submit their name, their value draw, a "bid" corresponding to the highest amount they would pay for a second eCoin, and an "ask" corresponding to the lowest amount they would accept to part with the eCoin they already own. Students keep their consumer and producer surpluses from each round as "extra credit" points. \code{equilibriumGame} tabulates the supply and demand schedules; calculates the equilibrium (with the help of a C++ helper function provided by "David" on Stack Overflow, \url{https://stackoverflow.com/questions/23830906/intersection-of-two-step-functions/}); graphs the equilibrium; and tabulates the scores for each student.
}
\references{
Holt, Charles A. (1996). Classroom games: Trading in a pit market. \emph{Journal of Economic Perspectives,} 10(1), pp.193-203.
}
