% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupList.R
\name{groupList}
\alias{groupList}
\title{Creates a random list of partner groups}
\usage{
groupList(studentList, n = 2)
}
\arguments{
\item{studentList}{(required) is an object containing the list of student names.}

\item{n}{is the size of each group.}
}
\value{
\code{studentList} returns list of names and group numbers. .
}
\description{
Creates a random list of partner groups
}
\details{
\code{partnerList} uses a random number to divide a \code{studentList} groups of size N.
}
