% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobbyGame.R
\name{lobbyGame}
\alias{lobbyGame}
\title{Tabulate results for an all-pay auction game.}
\usage{
lobbyGame(sheet, endowment = 5, prize = 4, seed = NULL, ...)
}
\arguments{
\item{sheet}{(required) is a character string sheet ID corresponding to the Google Sheet containing the individual submissions.}

\item{endowment}{is the size of the initial endowment of points the instructor wishes to give each student.}

\item{prize}{is the value of the license or prize each interest group (student) is bidding on.}
}
\value{
\code{type} returns the type of activity (allpayGame).

\code{results} returns the original submissions (with market prices and points per round added).

\code{winner} returns the name of the winner and the size of the prize.

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Tabulates and assigns points for the results of a simple in-class entry and exit game using random values.
}
\details{
\code{lobbyGame} tabulates the results of a simple lobbying game based on a simplified version of Goeree and Holt (1999). By default students receive an "endowment" of "extra credit" points (default is 5) at the beginning of a round of bidding. In the first iteration, students "bid" on a monopoly license "prize" (default is 4 points) by making an anonymous number of "contributions" to politicians in order to influence the decision by the government. Each contribution costs 1 point, and is tabulated as the number of times a student submits the Google Form. The winner of the auction is determined by a lottery from a random row sampled from the responses. In this iteration, students submit bids orally or electronically in real time according to an all-pay auction.
}
\references{
Goeree, Jacob K. & Charles A. Holt (1999). Classroom Games: Rent-Seeking and the Inefficiency of Non-market Allocations. \emph{Journal of Economic Perspectives,} 13(3), pp.217-226.
}
