% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipdGame.R
\name{multipdGame}
\alias{multipdGame}
\title{Tabulate results of a prisonor's dilemma game.}
\usage{
multipdGame(sheet, payoff = c(0, -0.5, 3, 1), ...)
}
\arguments{
\item{sheet}{(required) is a character string ID corresponding to the Google Sheet containing the individual submissions.}

\item{payoff}{(required) is an atomic vector indicating the interdependent payoffs from the different strategy pairs {Respondent, Class}: c({Compete, Compete}, {Collude, Compete}, {Compete, Collude}, {Collude, Collude})}
}
\value{
\code{type} returns the type of activity (multipdGame).

\code{equliibrium} returns the equilibrium strategy chosen by the majority of the class.

\code{payoff} returns the payoff matrix.

\code{results} returns the original submissions (with equilibria and points per round added).

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Tabulates and assigns points for the results of a simple in-class prisoner's dilemma game.
}
\details{
\code{multipdGame} tabulates the results of a simple prisoner's dilemma game in which students' points depend on their strategy and the strategy chosen by the majority of the rest of the class.
}
