% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_econGame.R
\name{plot.econGame}
\alias{plot.econGame}
\title{Plot econGame results}
\usage{
\method{plot}{econGame}(econGame, round = NULL, nrow = 1, ncol = NULL, binwidth = NULL, ...)
}
\arguments{
\item{econGame}{(required) is a list created by an \code{econGame} function.}

\item{round}{is the round of the game that you want to plot (defaults to NULL for all rounds).}

\item{nrow}{is the number of rows to use for \code{ggarrange} (defaults to 1).}

\item{ncol}{is the number of columns to use for \code{ggarrange} (defaults to the number of rounds).}

\item{binwidth}{is the number of bins to use for histograms.}

\item{...}{other optional graphical parameters.}
}
\value{
\code{plot.econGame} returns a \code{ggarrange} or \code{ggplot} object that can be plotted and exported.
}
\description{
Plots objects of class \code{econGame} created by tabulating the results of one of the games in the \code{econGame} package
}
\details{
\code{plot.econGame} currently supports the following games:
\code{equilibriumGame} for simple market equilibrium.
\code{entryGame} for a simple entry and exit game.
\code{staghuntGame} for a bar plot of the outcomes of the stag hunt game.
\code{cournotGame} for a bar plot of the outcomes of the Cournot duopoly game.
\code{stackelbergGame} for a bar plot of the outcomes of the Stackelberg duopoly game.
\code{bertrandGame} for a histogram of the price strategies and equlibrium prices for the Bertrand duopoly game.
\code{multipdGame} for a histogram of the strategies of a multi-player prisoner's dilemma game.
}
