% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutionGame.R
\name{pollutionGame}
\alias{pollutionGame}
\title{Tabulate pollution game results.}
\usage{
pollutionGame(sheet, price = 3, externality = 2, practice = TRUE, ...)
}
\arguments{
\item{sheet}{(required) is a sheet ID corresponding to the Google Sheets location containing the individual submissions.}

\item{price}{is the price of the output each "firm" produces (default is 3).}

\item{externality}{is the external cost of pollution for each unit of output firms produce (default is 2).}

\item{practice}{is a logical indicating whether round 1 is practice (default is \code{TRUE}).}
}
\value{
\code{type} returns the type of activity (pollutionGame).

\code{results} returns the original submissions (with market price and points added).

\code{schedules} returns a list containing the supply and demand schedules for each round.

\code{equilibria} returns a list containing the equilibria for each round.

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Tabulates and assigns points for the results of a four in-class pollution policy simulations using random cleanup-cost parameters. The four policies the results compare include: (1) no regulation; (2) command and control regulation, in which firms may only permit a fixed amount; (3) a pollution tax, in which firms must pay a tax equal to their external costs; and (4) cap and trade, in which firms receive free initial pollution permits and may trade in order to buy or sell additional permits.
}
\details{
\code{pollutionGame} tabulates the results of a pollution game based on Nugent (1997) and some modifications to it. At the beginning of the game, each "firm" (student) "draws" two independent random cleanup costs from a discrete uniform distribution through a link to a Google Sheet. The lower cleanup cost represents the cost of cleaning up the first unit of pollution and the higher number represents the cost of cleaning up the second unit of pollution (if applicable), so that cleanup follows the Law of Increasing Costs. In the default example, the cleanup costs are distributed on a uniform discrete distribution from 1 to 4. During each round of the game, firms choose how much to produce (0, 1, or 2 units) and how much pollution from their production to clean up (0, 1, or 2).
In the first round (which may be designated as practice to not count for points), there is no regulation. Students receive the price of their output as "extra credit" points (default is 3), internal production costs are set to zero (for simplicity), and impose a "social cost" for which they do not pay (default is 2). In this round, profit-maximizing firm would produce 2 units of output, pay no cost for either production or the externalities they impose, and earn a profit of 6 points.
In the second round, firms face command and control regulation that limits them to 1 unit of pollution. This means that students may produce the first unit of output at zero cost, but producing a second unit of output automatically requires them to clean up the pollution using and pay a cost equal to the lower cleanup cost parameter.
In the third round, firms face a pollution tax equal to any external costs they impose on society (by default equal to 2). In the default settings, the price (3) is greater than the externality (2) and any profit-maximizing firm chooses to produce both units of output. Firms must still choose how much to clean up: If a firm's \emph{lower} cleanup cost is less than the tax, then that firm will maximize profits by cleaning up the first unit of pollution; if their \emph{higher} cleanup cost is also less than the tax, then it will clean up the second unit as well.
In the fourth and final round, firms face a cap and trade mechanism that allocates one "free" permit-to-pollute to each firm (much like the command and control round), except this time firms may trade their pollution rights in a market. Firms cannot immediately choose their output in this round (although most would end up choosing 2 units). Instead, students must decide how much they would \emph{bid} to buy an additional permit and how much they would \emph{ask} to sell their "free" permit. The \code{pollutionGame} code calculates the optimal output and cleanup decisions, and computes the profits net of revenues or costs from permit trades.
}
\references{
Nugent, R. A. (1997). Teaching Tools: A Pollution Rights Trading Game \emph{Economic Inquiry,} 35(3), 679-685.
}
