% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMarketGameApp.R
\name{runMarketGameApp}
\alias{runMarketGameApp}
\title{Run a \code{marketGames} app.}
\usage{
runMarketGameApp()
}
\arguments{
\item{game}{(required) the name of the \code{marketGames} app you want to run, which can be one of:
\enumerate{
\item \code{equlibriumGame} to tabulate results for a simple supply-and-demand equilibrium game.
\item \code{entryGame} for a two-sector entry-and-exit game.
\item \code{bertrandGame} for a Bertrand duopoly game.
\item \code{cournotGame} for a Cournot duopoly game.
\item \code{stackelbergGame} for a sequential leader-follower game.
\item \code{multipdGame} for a multi-player prisoner's dilemma game.
\item \code{staghuntGame} for a two-person stag hunt gmae.
}}
}
\description{
This is a wrapper function for the game-specific \code{run_Game_App()} functions in the marketGames package.
}
