% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackelbergGame.R
\name{stackelbergGame}
\alias{stackelbergGame}
\title{Tabulate results for a simple in-class Stackelberg duopoly game.}
\usage{
stackelbergGame(
  resultsSheet = NULL,
  leaderSheet = NULL,
  a = 10,
  b = -1,
  c = 6,
  f = 0,
  partners = "random",
  names = NULL,
  ...
)
}
\arguments{
\item{a}{is the value of the intercept of the linear inverse-demand function (default is 10).}

\item{b}{is the value of the slope of the linear inverse-demand function (default is -1).}

\item{c}{is the value of the firm's marginal cost (default is 6).}

\item{f}{is the value of the firm's fixed cost (default is 0).}

\item{sheet}{(required) is a character string sheet ID corresponding to the Google Sheets location containing the individual submissions.}
}
\value{
\code{type} returns the type of activity (cournotGame).

\code{results} returns the original submissions (with equilibria and points per round added).

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Tabulate results for a simple in-class Stackelberg duopoly game.
}
\details{
\code{cournotGame} tabulates the results of a simple Stackelberg duopoly game in which students' points depend on their strategy and the strategy chosen by their randomly-assigned partners.
}
