% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultimatumGame.R
\name{ultimatumGame}
\alias{ultimatumGame}
\title{Tabulate ultimatum game results.}
\usage{
ultimatumGame(sheet, ...)
}
\arguments{
\item{sheet}{(required) is a character string url corresponding to the Google Sheets location containing the individual submissions.}
}
\value{
\code{type} returns the type of activity (ultimatumGame).

\code{results} returns the original submissions.

\code{grades} returns the aggregated, stacked points "won" by each student for the entire activity.
}
\description{
Tabulates and assigns points for the results of a simple in-class ultimatum game.
}
\details{
\code{ultimatumGame} tabulates the results of a simple ultimatum game based on Thaler (1988). Students form pairs and determine a "proposer" and a "responder" by flipping a coin (or some other fair mechanism). The pair receives 5 "extra credit" points to share. The proposer offers to give a number of points to the responder (between 0 and 5) and the responder decides whether to accept or reject. If the responder accepts, then the pair shares the points according to the aggreement; if the responder rejects, both players get nothing.
}
\references{
Holt, Charles A. (1996). Classroom games: Trading in a pit market. \emph{Journal of Economic Perspectives,} 10(1), pp.193-203.
}
